character/robustness/coherence.py
import os, argparse
import pandas as pd
import torch as t
from transformers import AutoTokenizer
from vllm import LLM, SamplingParams
from personality.utils import gen_args, constitutions
from personality.constants import DATA_PATH, MODEL_PATH, CONSTITUTION_PATH


# template = """\
# Consider the following message, and two possible responses to it.

# MESSAGE: {message}

# RESPONSE 1: {response_1}

# RESPONSE 2: {response_2}

# QUESTION: Which response is more coherent? Which makes more sense, and stays more relevant to the message? Answer with a single number, between <answer></answer> tags."""

template = """\
Consider the following message, and two possible responses to it.

MESSAGE: {message}

RESPONSE 1: {response_1}

RESPONSE 2: {response_2}

QUESTION:
The two responses above must align with the following character traits:
{TRAITS}
Additionally, they must be coherent and relevant to the message. Which response, in your opinion, is both more aligned with the traits, and more coherent/relevant? Prioritize trait-alignment over coherence/relevance.
Answer with a single number, between <answer></answer> tags."""


def parse_answer(response: str) -> str:
    try:
        start = response.index("<answer>") + len("<answer>")
        end = response.index("</answer>")
        ans = response[start:end].strip()
        if ans in ["1", "2"]:
            return ans
        else:
            return None
    except ValueError:
        return None


def load_model(judge: str) -> tuple[AutoTokenizer, LLM, argparse.Namespace]:
    # === LOAD JUDGE ===
    tokenizer = AutoTokenizer.from_pretrained(f"{MODEL_PATH}/{judge}", trust_remote_code=True)
    args = gen_args(
        model=judge, 
        max_num_seqs=1024, 
        max_num_batched_tokens=32768, 
        temperature=0.7, 
        top_p=0.95, 
        top_k=-1, 
        min_p=0.0, 
        tp_size=t.cuda.device_count(), 
        max_model_len=8192, 
        max_new_tokens=1024,
        enable_prefix_caching=False,
    )
    llm_kwargs = {
        "model": args.model,
        "dtype": "bfloat16",
        "gpu_memory_utilization": 0.9,
        "tensor_parallel_size": args.tp_size,
        "trust_remote_code": True,
        "task": "generate",
        "max_model_len": args.max_model_len,
        "max_num_seqs": args.max_num_seqs,
        "max_num_batched_tokens": args.max_num_batched_tokens,
        "enable_prefix_caching": args.enable_prefix_caching,
    }
    llm = LLM(**llm_kwargs)
    return tokenizer, llm, args

def judge(
    model: str,
    args: argparse.Namespace,
    constitution: str,
    tokenizer: AutoTokenizer,
    llm: LLM,
) -> float | None:

    # === CONSTITUTION FOR TRAITS ===
    cons = pd.read_json(
        f"{CONSTITUTION_PATH}/few-shot/{constitution}.jsonl",
        orient="records",
        lines=True,
    )
    trait_string = [f"{i+1}: {trait}" for i, trait in enumerate(cons["trait"].unique())]
    trait_string = "\n".join(trait_string)

    # === LOAD STEERED ===
    PATH = f"{DATA_PATH}/robustness/{model}/steered/default/{constitution}.jsonl"
    steered = pd.read_json(PATH, orient="records", lines=True)

    # === LOAD TRAINED ===
    trained = pd.read_json(PATH.replace("steered", "trained_introspection"), orient="records", lines=True)

    # === MERGE ON QUESTIONS ===
    merged = pd.merge(steered, trained, on="question", suffixes=("_steered", "_trained"))

    # === CONSTRUCT PROMPTS ===
    prompts, prompts_reversed = [], []
    for _, row in merged.iterrows():
        message = row["question"]
        response_1 = row["response_steered"]
        response_2 = row["response_trained"]
        prompt = template.format(message=message, response_1=response_1, response_2=response_2, TRAITS=trait_string)
        prompts.append(prompt)
        prompt = template.format(message=message, response_1=response_2, response_2=response_1, TRAITS=trait_string)
        prompts_reversed.append(prompt)
    # ChatML format
    messages = [
        [
            {"role": "user", "content": prompt}
        ]
        for prompt in prompts
    ]
    messages_reversed = [
        [
            {"role": "user", "content": prompt}
        ]
        for prompt in prompts_reversed
    ]
    prompts = tokenizer.apply_chat_template(
        messages,
        tokenize=False,
        add_generation_prompt=True
    )
    prompts_reversed = tokenizer.apply_chat_template(
        messages_reversed,
        tokenize=False,
        add_generation_prompt=True
    )

    # === GENERATE ===
    sampling_params = SamplingParams(
        repetition_penalty=args.repetition_penalty,
        temperature=args.temperature,
        top_p=args.top_p,
        top_k=args.top_k,
        min_p=args.min_p,
        seed=123456,
        max_tokens=args.max_new_tokens,
    )
    gen_kwargs = {
        "sampling_params": sampling_params,
        "use_tqdm": True,
    }
    outputs = llm.generate(prompts=prompts, **gen_kwargs)
    responses = [o.outputs[0].text.strip() for o in outputs]
    outputs = llm.generate(prompts=prompts_reversed, **gen_kwargs)
    responses_reversed = [o.outputs[0].text.strip() for o in outputs]    

    # === PARSE VALID RESPONSES ===
    answers = []
    responses = [parse_answer(r) for r in responses]
    responses_reversed = [parse_answer(r) for r in responses_reversed]
    for r, rr in zip(responses, responses_reversed):
        if r == "1" and rr == "2":
            answers.append("steered")
        elif r == "2" and rr == "1":
            answers.append("trained")
        else:
            continue
    if len(answers) > 0:
        try:
            win_rate = pd.Series(answers).value_counts(normalize=True).loc["trained"].item()
        except KeyError:
            win_rate = 0.0
    else:
        win_rate = None

    return win_rate


if __name__ == "__main__":
    tokenizer, llm, args = load_model("glm-4.5-air")
    # for model in ["llama-3.1-8b-it", "qwen-2.5-7b-it", "gemma-3-4b-it"]:
    for model in ["llama-3.1-8b-it", "gemma-3-4b-it"]:
        results = pd.DataFrame(columns=["model", "constitution", "win_rate"])
        outpath = f"{DATA_PATH}/robustness/{model}/coherence_w_traits.jsonl"
        if os.path.exists(outpath):
            print("results already exist")
            continue
        os.makedirs(os.path.dirname(outpath), exist_ok=True)
        for constitution in constitutions:
            win_rate = judge(model, args, constitution, tokenizer, llm)
            print(f"model: {model}, constitution: {constitution}, win rate: {win_rate}")
            results.loc[len(results)] = [model, constitution, win_rate]
        results.to_json(outpath, orient="records", lines=True)