requirements.txt
accelerate
bitsandbytes
datasets
deepspeed
einops
evaluate
ipykernel
ipywidgets
isort
jsonlines
loralib
matplotlib
optimum
packaging
peft
scikit-learn
tensorboard
torchmetrics
tqdm
transformers
transformers_stream_generator
wandb
wheel